patch: move clients to subfolder
