+++ src/jdom/bus.ts
@@ -71,7 +71,7 @@ import RealTimeClockServer from "../servers/realtimeclockserver"
 import { SRV_ROLE_MANAGER } from "../../src/jdom/constants"
 import { JDTransport } from "./transport/transport"
 import { BusStatsMonitor } from "./busstats"
-import { RoleManagerClient } from "./rolemanagerclient"
+import RoleManagerClient from "./clients/rolemanagerclient"
 import JDBridge from "./bridge"
 import IFrameBridgeClient from "./iframebridgeclient"
 import { randomDeviceId } from "./random"
