+++ src/jdom/clients/settingsclient.ts
@@ -1,10 +1,10 @@
-import { CHANGE, SettingsCmd } from "./constants"
-import { jdpack, jdunpack } from "./pack"
-import Packet from "./packet"
-import { InPipeReader } from "./pipes"
-import { JDService } from "./service"
-import { JDServiceClient } from "./serviceclient"
-import { bufferToString, stringToBuffer } from "./utils"
+import { CHANGE, SettingsCmd } from "../constants"
+import { jdpack, jdunpack } from "../pack"
+import Packet from "../packet"
+import { InPipeReader } from "../pipes"
+import { JDService } from "../service"
+import { JDServiceClient } from "../serviceclient"
+import { bufferToString, stringToBuffer } from "../utils"
 
 export default class SettingsClient extends JDServiceClient {
     constructor(service: JDService) {
