+++ src/jdom/jacdac-jdom.ts
@@ -22,8 +22,7 @@ export * from "./speedtest"
 export * from "./spec"
 export * from "./client"
 export * from "./serviceclient"
-export * from "./sensoraggregatorclient"
-export * from "./modelrunner"
+export * from "./clients/sensoraggregatorclient"
 export * from "./iframeclient"
 export * from "./iframebridgeclient"
 export * from "./frame"
@@ -32,9 +31,7 @@ export * from "./command"
 export * from "./packetfilter"
 export * from "./tracerecorder"
 export * from "./traceview"
-export * from "./traceplayer"
 export * from "./flags"
-export * from "./settingsclient"
 export * from "./pack"
 export * from "./packobject"
 export * from "./light"
@@ -44,9 +41,11 @@ export * from "./serviceprovider"
 export * from "./controlserver"
 export * from "./color"
 export * from "./makecode"
-export * from "./rolemanagerclient"
 export * from "./qualityofservice"
 export * from "./bridge"
+export * from "./clients/modelrunnerclient"
+export * from "./clients/settingsclient"
+export * from "./clients/rolemanagerclient"
 export * from "./transport/transport"
 export * from "./transport/bluetooth"
 export * from "./transport/usbio"
